Move FixtureSet::ClassCache to seaprate file
